+++ _PROJECT/model/employee.py
@@ -1,6 +1,7 @@
 class Employee:
     def __init__(self, kennitala:int=0, name:str="", role:str="", address:str="", phone_number:int=0, email:str=""):
         self.kennitala = kennitala
+        self.print_kennitala = kennitala[:6] + "-" + kennitala[6:]
         self.name = name
         self.role = role
         self.address = address
