+++ _PROJECT/ui/employee_ui.py
@@ -1,15 +1,17 @@
 from logic.employee_logic import EmployeeLogic
 from model.employee import Employee
-from funclibrary.functions import clear_screen
+from utils.utils import UIUtils
 
 
 class EmployeeUI:
     def __init__(self, wrapper) -> None:
        self.logic_wrapper = wrapper
+       self.ui_utils = UIUtils()
 
     def menu_output(self):
         header = f"[EMPLOYEES]"
         
+        print()
         print(header)
         print()
         print(f"1. List all employees")
@@ -22,15 +24,19 @@ class EmployeeUI:
         """ Takes in input from user """
         while True:
             self.menu_output()
-            user_input = input("Enter your choice: ").lower()
-            clear_screen()
+            user_input = input("\nEnter your choice: ").lower()
+            # Clears screen
+            # self.ui_utils.clear_screen()
             if user_input == "q":
                 print("Quitting")
                 break
             elif user_input == "1":
                 employees = self.logic_wrapper.get_all_employees()
                 employees = self.logic_wrapper.get_all_employees()
+                print()
+                print("[ALL EMPLOYEES]")
+                print()
                 for index, employee in enumerate(employees):
-                    print(f"{index+1}:\t{employee.name}")
+                    print(f"{index+1:>2}.{' name: ':^2}{employee.name:<}, {employee.role}\n      {'kt: ' + employee.print_kennitala}")
             else:
                 print("Invalid")
