+++ _PROJECT/ui/main_menu_ui.py
@@ -1,7 +1,7 @@
-from logic.logic_wapper import LogicWrapper
+from logic.logic_wrapper import LogicWrapper
 from ui.employee_ui import EmployeeUI
 from ui.voyage_ui import VoyageUI
-from funclibrary.functions import clear_screen
+#from funclibrary.functions import clear_screen
 
 class MainMenuUI:
     def __init__(self) -> None:
@@ -10,6 +10,7 @@ class MainMenuUI:
     def menu_output(self):
         header = f"[MAIN]"
         
+        print()
         print(header)
         print()
         print(f"1. Employees")
@@ -19,8 +20,8 @@ class MainMenuUI:
     def input_prompt(self):
         while True:
             self.menu_output()
-            user_input = input("Enter your choice: ").lower()
-            clear_screen()
+            user_input = input("\nEnter your choice: ").lower()
+            #clear_screen()
             if user_input == "q":
                 print("Quitting")
                 break
